Move icon handlers to root window
