+++ src/ipc/channels.ts
@@ -4,7 +4,7 @@ type ChannelToArgsMap = {
   'redux/get-initial-state': () => unknown;
   'redux/action-dispatched': (action: AnyAction) => void;
   'servers/fetch-info': (urlHref: string) => [urlHref: string, version: string];
-  'notifications/get-icon': (urlHref: string) => string;
+  'notifications/fetch-icon': (urlHref: string) => string;
 };
 
 export type Channel = keyof ChannelToArgsMap;
