+++ src/rootWindow.ts
@@ -5,10 +5,7 @@ import { setupRendererErrorHandling } from './errors';
 import { setupI18n } from './i18n/renderer';
 import { createRendererReduxStore } from './store';
 import { App } from './ui/components/App';
-import { setupRootWindowIcon } from './ui/rootWindow/icon';
 import { whenReady } from './whenReady';
-import './notifications/renderer';
-import './servers/renderer';
 
 const start = async (): Promise<void> => {
   const reduxStore = await createRendererReduxStore();
@@ -17,7 +14,11 @@ const start = async (): Promise<void> => {
 
   setupRendererErrorHandling('rootWindow');
   await setupI18n();
-  setupRootWindowIcon();
+
+  await Promise.all([
+    import('./notifications/renderer'),
+    import('./servers/renderer'),
+  ]);
 
   const container = document.getElementById('root');
